server-http-python-lambda/server/requirements.txt
boto3==1.38.38
awslabs.mcp_lambda_handler==0.1.6